Fix (s32) pointer truncation in print.c alignment loops
print_getLettersFromFont uses while((s32)chunkDataPtr % 8) for
alignment, truncating 64-bit pointers to 32 bits. Changed to
(uintptr_t). Also fix same pattern in code_C0E60.c texture load.

+++ src/core2/code_C0E60.c
@@ -190,7 +190,7 @@ void func_80348044(Gfx **gfx, BKSprite* sprite, s32 frame, s32 tmem, s32 rtile, 
 
     *width = (s32) texture_block->w;
     *height = (s32) texture_block->h;
-    for(timg = (s32)(texture_block + 1); timg % 8; timg++);
+    for(timg = (uintptr_t)(texture_block + 1); timg % 8; timg++);
 
     if (sp144 == 0) {
         D_8038608C = *width;
